Add an ideas links section
Imagine using git (the CLI software, not an online repo, like github) as an offline-first database "backend", a graph daytabase, for a peer-to-peer frontend that is structured like a wiki that uses hash keys to prove expertise in-field/in-topic for greater weight in expert posts and greater weight for expert posts over public posts. Accurate posts would float! "Cooker" rubbish would sink without a trace.